Fix window opacity bug (#1283)
With the tab refactor, we introduced a regression that prevented window
transparency from working. Rather than applying overrides of the app
background color (where we record window opacity) to the body of the
document, we were applying it to the main div. This meant that while the
main div was transparent, the body was not (since the default background
color we set for the body is opaque). With this change, we update the
body background directly.