Adding install instructions to the docsite (#1312)
This updates the platform toggle to include Windows as well and to use
the UAParser library to extract the OS from the UserAgent. This also
improves the legibility of the platform selector pill and adds a new
PlatformItem component that can be used to specify platform-specific
content.

This also adds a new Getting Started document with installation
instructions and links to the different package manager listings.