Use Zig for Windows and ARM64 Linux (#1713)
I'm making steps to simplify our build dependencies, consolidating our C
compiler requirements so we only depend on Zig.

Before, we used Zig for x64 but not for arm64. This meant that users
using an ARM dev machine would need to install `build-essentials` and
Zig. We also required MinGW-w64 on Windows, which is a pain to install
since it can be provided by a bunch of different tools, the smallest of
which is like a 2GB install.